tech(api): move center repository in organization-entities context
