+++ api/tests/certification/configuration/integration/infrastructure/repositories/center-repository_test.js
@@ -1,5 +1,5 @@
 import * as centerRepository from '../../../../../../src/certification/configuration/infrastructure/repositories/center-repository.js';
-import { CenterTypes } from '../../../../../../src/certification/enrolment/domain/models/CenterTypes.js';
+import { CenterTypes } from '../../../../../../src/organizational-entities/domain/read-models/CenterTypes.js';
 import { expect } from '../../../../../test-helper.js';
 import { databaseBuilder, knex } from '../../../../../tooling/databases.js';
 import { domainBuilder } from '../../../../../tooling/domain-builder/domain-builder.js';
