+++ api/tests/organizational-entities/integration/domain/usecases/update-certification-center.usecase.test.js
@@ -1,6 +1,6 @@
-import * as centerRepository from '../../../../../src/certification/enrolment/infrastructure/repositories/center-repository.js';
 import { CenterForAdmin } from '../../../../../src/organizational-entities/domain/models/CenterForAdmin.js';
 import { updateCertificationCenter } from '../../../../../src/organizational-entities/domain/usecases/update-certification-center.usecase.js';
+import * as centerRepository from '../../../../../src/organizational-entities/infrastructure/repositories/center-repository.js';
 import * as certificationCenterForAdminRepository from '../../../../../src/organizational-entities/infrastructure/repositories/certification-center-for-admin.repository.js';
 import * as complementaryCertificationHabilitationRepository from '../../../../../src/organizational-entities/infrastructure/repositories/complementary-certification-habilitation.repository.js';
 import * as dataProtectionOfficerRepository from '../../../../../src/organizational-entities/infrastructure/repositories/data-protection-officer.repository.js';
